kvx-fast: work around FAST NMS perfmance degradation
Summary:
In one of the versions of the Kalray KVX compiler, a part of the
FAST OpenCL code was impacted by a huge performance loss.

Until the precise cause is understood, rewrite this part of the
code in a way that is not affected.

ref T16577

Test Plan:
CI tests should pass and benches should show performances
similar to older compiler versions.

Reviewers: mkapfer, mqho, etournier, O72 kaf!, O78 kstuff!

Maniphest Tasks: T16577

Differential Revision: https://phab.kalray.eu/D13284